Recover BWC test suite, but keep it disabled.

+++ integ-test/src/test/java/org/opensearch/sql/legacy/SQLIntegTestCase.java
@@ -139,8 +139,10 @@ public abstract class SQLIntegTestCase extends OpenSearchSQLRestTestCase {
    */
   @AfterClass
   public static void cleanUpIndices() throws IOException {
-    wipeAllOpenSearchIndices();
-    wipeAllClusterSettings();
+    if (System.getProperty("tests.rest.bwcsuite") == null) {
+      wipeAllOpenSearchIndices();
+      wipeAllClusterSettings();
+    }
   }
 
   protected void setQuerySizeLimit(Integer limit) throws IOException {
